refactor: move fetcher settings components to separate directory
