+++ src/routes/project/[projectId]/settings/slr/+page.svelte
@@ -2,7 +2,7 @@
     import { goto } from "$app/navigation";
     import ProjectSettingsLayout from "$lib/components/composites/settings/project-settings/ProjectSettingsLayout.svelte";
     import MaybeAsDecisionSetting from "$lib/components/composites/settings/project-settings/slr/MaybeAsDecisionSetting.svelte";
-    import FetcherSettings from "$lib/components/composites/settings/project-settings/slr/FetcherSettings.svelte";
+    import FetcherSettings from "$lib/components/composites/settings/project-settings/slr/fetcher/FetcherSettings.svelte";
     import { ProjectStatus } from "$lib/model/api/project.js";
     import { resource } from "$lib/resource.svelte";
     import { isCurrentUserProjectAdmin } from "../helper";
